% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agencies.R
\name{build_regextable}
\alias{build_regextable}
\title{Build agency regex lookup table}
\usage{
build_regextable(sheet_url)
}
\arguments{
\item{sheet_url}{Google Sheets URL for the agency crosswalk}
}
\value{
A list with a regex table and duplicate aliases
}
\description{
Build agency regex lookup table
}
